Move display implementation into `widgets::display::DisplayUI`
